ref(Movies): useWidth with debounce
