+++ src/utils/hooks/useWidth.js
@@ -1,0 +1,21 @@
+import { useEffect, useState } from 'react';
+
+const useWidth = () => {
+  const [width, setWidth] = useState(window.innerWidth);
+
+  useEffect(() => {
+    let timeout;
+    const resizeListener = () => {
+      clearTimeout(timeout);
+      timeout = setTimeout(() => setWidth(window.innerWidth), 200);
+    };
+    window.addEventListener('resize', resizeListener);
+    return () => {
+      window.removeEventListener('resize', resizeListener);
+    };
+  }, []);
+
+  return width;
+};
+
+export default useWidth;
